misc fixes in r10d
1. Remove hack for server where android-13,15,17 were missing
2. Remove --disable-libquadmath
3. Fixed build-gnu-libstdc++.sh thumb build where
   "-ffunction-sections -fdata-sections" were dropped
4. Fixed llvm clang3.5/windows build where mingw-64 complains about -fPIC
5. Fixed bashism in make-standalone-toolchain.sh
6. Fixed "package-release.sh --separate-64"
7. Relaxed patch-sources.sh against non-existance patch directory
8. Refresh build/tools/toolchain-patches/gcc/0010-Fix-PR-target-63209.patch
9. Fixed warning about unused vars in cpu-features.c
10. Fixed sources/android/support/Android.mk about -fvisibility-inlines-hidden
   is only foo C++
11. Fixed tests/build/check-armeabi-v7a-prebuilts/build.sh on MacOSX 10.6.8
12. Fixed tests/build/issue64679-prctl/build.sh to include GLES2/gl2ext.h
   which can be compiled alone now
13. Excluded tests/device/b8708181-Vector4/jni/Vector4.cpp from clang3.4
14. Removed some comments in tests/device/test-libc++-*-full/jni/Android.mk
14. misc run-test.sh fixes

+++ tests/run-tests.sh
@@ -570,11 +570,19 @@ is_incompatible_abi ()
         local APP_ABIS=`get_build_var $PROJECT APP_ABI`
         APP_ABIS=$APP_ABIS" "
         if [ "$APP_ABIS" != "${APP_ABIS%%all*}" ] ; then
-        # replace the first "all" with all available ABIs
-          ALL_ABIS=`get_build_var $PROJECT NDK_ALL_ABIS`
-          APP_ABIS_FRONT="${APP_ABIS%%all*}"
-          APP_ABIS_BACK="${APP_ABIS#*all}"
-          APP_ABIS="${APP_ABIS_FRONT}${ALL_ABIS}${APP_ABIS_BACK}"
+        # replace "all", "all32" and "all64"
+          _EXPANDED=`get_build_var $PROJECT NDK_APP_ABI_ALL_EXPANDED`
+          _FRONT="${APP_ABIS%%all*}"
+          _BACK="${APP_ABIS#*all}"
+          APP_ABIS="${_FRONT}${_EXPANDED}${_BACK}"
+          _EXPANDED=`get_build_var $PROJECT NDK_APP_ABI_ALL32_EXPANDED`
+          _FRONT="${APP_ABIS%%all32*}"
+          _BACK="${APP_ABIS#*all32}"
+          APP_ABIS="${_FRONT}${_EXPANDED}${_BACK}"
+          _EXPANDED=`get_build_var $PROJECT NDK_APP_ABI_ALL64_EXPANDED`
+          _FRONT="${APP_ABIS%%all64*}"
+          _BACK="${APP_ABIS#*all64}"
+          APP_ABIS="${_FRONT}${_EXPANDED}${_BACK}"
         fi
         if [ "$APP_ABIS" = "${APP_ABIS%$ABI *}" ] ; then
             echo "Skipping `basename $PROJECT`: incompatible ABI, needs $APP_ABIS"
@@ -911,9 +919,11 @@ if is_testable device; then
                 dump "   ---> TEST FAILED!!"
             fi
             adb_var_shell_cmd "$DEVICE" "" "rm -f $DSTPATH"
-            for DATA in $(ls $DATAPATHS); do
-                adb_var_shell_cmd "$DEVICE" "" "rm -f $DSTDIR/`basename $DATA`"
-            done
+            if [ -n "$DATAPATHS" ]; then
+                for DATA in $(ls $DATAPATHS); do
+                    adb_var_shell_cmd "$DEVICE" "" "rm -f $DSTDIR/`basename $DATA`"
+                done
+            fi
         done
         # Cleanup
         adb_var_shell_cmd "$DEVICE" "" rm -r $DSTDIR
@@ -971,7 +981,7 @@ if is_testable device; then
             adb_var_shell_cmd "$DEVICE" CPU_ABI2 getprop ro.product.cpu.abi2
             CPU_ABIS="$CPU_ABI1,$CPU_ABI2"
             CPU_ABIS=$(commas_to_spaces $CPU_ABIS)
-            if [ "$_NDK_TESTING_ALL_" = "yes" ]; then
+            if [ -n "$_NDK_TESTING_ALL_" ]; then
                 if [ "$CPU_ABI1" = "armeabi-v7a" -o "$CPU_ABI2" = "armeabi-v7a" ]; then
                     CPU_ABIS="$CPU_ABIS armeabi-v7a-hard"
                 fi
